style: correct layout logic
